% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_observation.R
\name{cluster_observation}
\alias{cluster_observation}
\title{Cluster observations}
\usage{
cluster_observation(conn, status, max_dist = 336, reset = FALSE)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{status}{highest status to include while clustering the observations.}

\item{max_dist}{maximum clustering distance in m.}

\item{reset}{reset the current clustering.
Defaults to \code{FALSE}}
}
\description{
Cluster observations
}
