% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to a database}
\usage{
connect_db(db = tempfile("territoria", fileext = ".sqlite"))
}
\arguments{
\item{db}{The file name of the database}
}
\description{
Connect to a database
}
