% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Calculate the distance matrix}
\usage{
distance_matrix(conn, max_dist = 336)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{max_dist}{maximum clustering distance in m.}
}
\description{
Calculate the distance matrix
}
