% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_distribution.R
\name{edge_distribution}
\alias{edge_distribution}
\title{Edge distribution}
\usage{
edge_distribution(conn, max_dist = 336)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{max_dist}{maximum clustering distance in m.}
}
\description{
This function returns the length of the edges of the Delaunay triangulation
per survey.
The function returns only edges with length smaller than twice \code{max_dist}.
}
