% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster.R
\name{get_cluster}
\alias{get_cluster}
\title{Get the information from the clusters}
\usage{
get_cluster(conn)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}
}
\description{
Get the information from the clusters
}
