% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total.R
\name{get_total}
\alias{get_total}
\title{Get the total number of individuals per region
Returns the sum of \code{value} per region.
Clusters spanning more than one region get each a fraction of \code{value} based
on the number of observations of the cluster in a region divided by the total
number of observations in the cluster.}
\usage{
get_total(conn, value = "IIF(max_status >= 2, 1, iif(n_total > 1, 0.5, 0))")
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{value}{A string representing the value per cluster to be calculated.
Default is \code{"IIF(max_status >= 2, 1, iif(n_total > 1, 0.5, 0))"}.
Available variables are
\itemize{
\item \code{max_status}: the highest status of a cluster.
\item \code{n_total}: the total number of observations per cluster
\item \code{n_region}: the number of observations in a cluster from the same region.
}}
}
\description{
Get the total number of individuals per region
Returns the sum of \code{value} per region.
Clusters spanning more than one region get each a fraction of \code{value} based
on the number of observations of the cluster in a region divided by the total
number of observations in the cluster.
}
