% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlikely.R
\name{get_unlikely_observation}
\alias{get_unlikely_observation}
\title{Get the unlikely observations}
\usage{
get_unlikely_observation(conn, crs = 31370)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{crs}{coordinate reference system to be assigned; object of class \code{crs}}
}
\description{
Get the unlikely observations
}
