% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_observations.R
\name{simulate_observations}
\alias{simulate_observations}
\title{Simulate a dataset in a square area}
\usage{
simulate_observations(
  density = 1e-05,
  area = 4e+06,
  gamma = 0.5,
  radius = 336,
  n_survey = 4,
  p_detection = 0.6,
  status_distribution = c(0.2, 0.5, 0.3)
)
}
\arguments{
\item{density}{Density as the number of territoria per m².}

\item{area}{Area in which to simulate territoria in m².}

\item{gamma}{interaction parameter of the Strauss process.
See \code{gamma} in \code{spatstat.core::rStrauss()}.}

\item{radius}{interaction radius in m.}

\item{n_survey}{Number of surveys.}

\item{p_detection}{Probability of detection within a survey.}

\item{status_distribution}{a weighting vector for statuses.
The order of the vector is the number of the status.}
}
\description{
Simulate a dataset in a square area
}
