% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlikely_status.R
\name{unlikely_status}
\alias{unlikely_status}
\title{Detect surveys with an unlikely status distribution}
\usage{
unlikely_status(conn, threshold = 0.5, status_split = 2, alpha = 0.01)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{threshold}{A numeric value between 0 and 1 indicating the average
proportion of the observations with status above or equal to the
\code{status_split }.
Default of 0.5.}

\item{status_split}{An integer values indicating which status level splits
the status into two groups.
The default is 2.}

\item{alpha}{A numeric value between 0 and 1 indicating the family-wise Type
I error.
Default of 0.01.}
}
\value{
A numeric value between 0 and 1 indicating the proportion of
surveys with an unlikely status distribution.
The list of unlikely surveys is written to the database.
}
\description{
Detect surveys with an unlikely status distribution
}
