% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlikely_survey_area.R
\name{unlikely_survey_area}
\alias{unlikely_survey_area}
\title{Detect surveys with an unlikely surveyed area}
\usage{
unlikely_survey_area(
  conn,
  relevant_area,
  max_area = 3e+06,
  dist = 50,
  crs = 31370,
  ratio = 0.95
)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{max_area}{the maximum area of the concave hull.}

\item{dist}{numeric or object of class \code{units}; buffer distance(s) for all, or for each of the elements in \code{x}.
In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{TRUE}, a numeric
\code{dist} is taken as distance in meters and a \code{units} object in \code{dist} is converted to meters.
In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{FALSE}, a numeric
\code{dist} is taken as degrees, and a \code{units} object in \code{dist} is converted to \code{arc_degree} (and warnings are issued).
In case \code{x} does not have geodetic coordinates (projected) then
numeric \code{dist} is assumed to have the units of the coordinates, and a \code{units} \code{dist} is converted to those if \code{st_crs(x)} is not \code{NA}.}

\item{crs}{coordinate reference system to be assigned; object of class \code{crs}}

\item{ratio}{numeric; fraction convex: 1 returns the convex hulls, 0 maximally concave hulls}
}
\description{
First we create a Delaunay triangulation of the survey area.
Then we ignore all triangles with an edge larger than \code{max_edge}.
Finally we calculate the area of the remaining triangles.
When the sum of the area of the remaining triangles is larger than
\code{max_area}, we consider the survey area unlikely.
}
