% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unlikely.R
\name{get_unlikely_summary}
\alias{get_unlikely_summary}
\title{Get an overview of the unlikely reasons for each survey and region}
\usage{
get_unlikely_summary(conn = conn)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}
}
\description{
Get an overview of the unlikely reasons for each survey and region
}
