% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlikely_edge_distribution.R
\name{unlikely_edge_distribution}
\alias{unlikely_edge_distribution}
\title{Check for unlikely distribution of Delaunay edges}
\usage{
unlikely_edge_distribution(conn, conn_reference, max_dist = 336, alpha = 0.01)
}
\arguments{
\item{conn}{a DBI connection to an SQLite database.}

\item{conn_reference}{a connection to the reference database.}

\item{max_dist}{maximum clustering distance in m.}

\item{alpha}{A numeric value between 0 and 1 indicating the family-wise Type
I error.
Default of 0.01.}
}
\description{
This function compares the distribution of Delaunay edges per survey in the
\code{conn} database with the distribution of Delaunay edges of all surveys in the
\code{conn_reference} database.
It uses a Wilcoxon test to check if the distribution of edges is
significantly different.
The function returns the proportion of surveys with an unlikely distribution
and writes the list of unlikely surveys to the database.
}
